+++ documentation/playground/src/atoms/TextFieldExamples.tsx
@@ -15,7 +15,16 @@ export function TextFieldExamples() {
           charCounter
           maxLength={100}
         />
-        <TextField label="Uncontrolled" required id="textfield-2" description="Help text" charCounter maxLength={200} />
+        <TextField
+          label="Uncontrolled"
+          required
+          addon={{ type: 'icon', placement: 'left', content: 'alarm-bell' }}
+          id="textfield-2"
+          description="Help text"
+          charCounter
+          characterCountText="Je kan nog %left% karakters typen."
+          maxLength={200}
+        />
         <TextField label="Controlled - no onChange" id="textfield-3" value="hello" charCounter maxLength={100} />
         <TextField label="Error" id="textfield-4" description="There is an error" state="error" type="password" />
         <TextField label="Small success" id="textfield-5" description="Ok!" state="success" size="small" />