+++ packages/antwerp-ui/react-components/src/lib/atoms/input/text-field/TextField.spec.tsx
@@ -53,25 +53,25 @@ describe('UI Components - Atoms - Input - TextField', () => {
   });
 
   it('should add an icon when iconLeft is defined', () => {
-    const { baseElement } = render(<TextField iconLeft="ai-single-neutral" />);
+    const { baseElement } = render(<TextField icon={{ name: 'ai-single-neutral', placement: 'left' }} />);
     expect(baseElement.getElementsByClassName('has-icon-left').length).toBe(1);
     expect(baseElement.getElementsByClassName('ai-single-neutral').length).toBe(1);
   });
 
   it('should add an icon when iconRight is defined', () => {
-    const { baseElement } = render(<TextField iconRight="ai-single-neutral" />);
+    const { baseElement } = render(<TextField icon={{ name: 'ai-single-neutral', placement: 'right' }} />);
     expect(baseElement.getElementsByClassName('has-icon-right').length).toBe(1);
     expect(baseElement.getElementsByClassName('ai-single-neutral').length).toBe(1);
   });
 
   it('should add an addon when addonLeft is defined', () => {
-    const { baseElement } = render(<TextField addonLeft="left" />);
+    const { baseElement } = render(<TextField addon={{ label: 'left', placement: 'left' }} />);
     expect(baseElement.getElementsByClassName('has-addon-left').length).toBe(1);
     expect(baseElement.getElementsByClassName('a-input__addon')[0].textContent === 'left').toBeTruthy();
   });
 
   it('should add ann addon when addonRight is defined', () => {
-    const { baseElement } = render(<TextField addonRight="right" />);
+    const { baseElement } = render(<TextField addon={{ label: 'right', placement: 'left' }} />);
     expect(baseElement.getElementsByClassName('a-input__addon').length).toBe(1);
     expect(baseElement.getElementsByClassName('has-addon-right').length).toBe(1);
   });