+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Core branding update to 6.2.3
 - Conditionally remove `aria-labbeledby` on the `ProgressBar` when an `aria-label` is used
+- Small improvements on `TextField` and `TextArea`: `iconLeft`, `iconRight`, `addonLeft` and `addonRight` are replaced by `addon` object
+
+### Added
+
+- `characterCountText` property on `TextField` and `TextArea`
 
 ## [7.0.0-beta.0] - 2023-03-30